Move ScheduledTaskForm to its own file
