+++ functionary/ui/forms/__init__.py
@@ -1,0 +1,2 @@
+from .scheduled_task import ScheduledTaskForm  # noqa
+from .tasks import TaskParameterForm  # noqa
